tiding up scripts and passing them to where they are used
